requirements.txt
Django==5.2.8
djangorestframework==3.14.0
celery==5.3.1
redis==4.5.4
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
transformers==4.35.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
newspaper3k==0.2.8
feedparser==6.0.12
whitenoise==6.6.0
dj-database-url==2.1.0
Pillow==10.2.0
lxml==5.1.0
google-genai
pytz==2023.3
# OCR libraries for image processing
Pillow==10.1.0
pytesseract==0.3.10